+++ application/ui/src/setup-tests.ts
@@ -34,3 +34,10 @@ Object.defineProperty(global, 'Request', {
     writable: false,
     value: RequestPolyfill,
 });
+
+// Mock ResizeObserver which is not available in jsdom
+global.ResizeObserver = vi.fn().mockImplementation(() => ({
+    observe: vi.fn(),
+    unobserve: vi.fn(),
+    disconnect: vi.fn(),
+}));
