+++ application/ui/src/shared/annotator/annotation-actions-provider.component.tsx
@@ -46,7 +46,7 @@ interface AnnotationsContextValue {
     annotations: Annotation[];
     addAnnotations: (shapes: Shape[], labels: Label[]) => void;
     deleteAnnotations: (annotationIds: string[]) => void;
-    updateAnnotations: (updatedAnnotations: Annotation[]) => void;
+    updateAnnotations: (updatedAnnotations: Annotation[], labels?: Label[]) => void;
     submitAnnotations: () => Promise<void>;
     isUserReviewed: boolean;
     isSaving: boolean;
@@ -86,12 +86,20 @@ export const AnnotationActionsProvider = ({
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [initialAnnotationsDTO, project?.task?.labels]);
 
-    const updateAnnotations = (updatedAnnotations: Annotation[]) => {
-        const updatedMap = new Map(updatedAnnotations.map((annotation) => [annotation.id, annotation]));
-
-        setAnnotations((prevAnnotations) =>
-            prevAnnotations.map((annotation) => updatedMap.get(annotation.id) ?? annotation)
-        );
+    const updateAnnotations = (updatedAnnotations: Annotation[], labels?: Label[]) => {
+        if (labels !== undefined) {
+            const idsToUpdate = new Set(updatedAnnotations.map((a) => a.id));
+            setAnnotations((prevAnnotations) =>
+                prevAnnotations.map((annotation) =>
+                    idsToUpdate.has(annotation.id) ? { ...annotation, labels } : annotation
+                )
+            );
+        } else {
+            const updatedMap = new Map(updatedAnnotations.map((annotation) => [annotation.id, annotation]));
+            setAnnotations((prevAnnotations) =>
+                prevAnnotations.map((annotation) => updatedMap.get(annotation.id) ?? annotation)
+            );
+        }
     };
 
     const addAnnotations = (shapes: Shape[], labels: Label[]) => {
